Added ther release plugin

+++ pet-clinic-data/src/main/java/com/charlie/petclinic/services/PetService.java
@@ -5,7 +5,7 @@ import com.charlie.petclinic.model.Pet;
 import java.util.Set;
 
 public interface PetService {
-    Pet findById(Long id);
+    Pet findById(Integer id);
     Pet save(Pet owner);
     Set<Pet> findAll();
 }
