+++ pet-clinic-data/src/main/java/com/charlie/petclinic/services/VetService.java
@@ -5,7 +5,7 @@ import com.charlie.petclinic.model.Vet;
 import java.util.Set;
 
 public interface VetService {
-    Vet findById(Long id);
+    Vet findById(Integer id);
     Vet save(Vet Vet);
     Set<Vet> findAll();
 }
